requirements.txt
airflow[postgres,crypto]
pyopenssl
psycopg2
bs4
absl-py